Split enrichment and init progress into smaller modules
